+++ src/commands/audio/stop.ts
@@ -1,4 +1,4 @@
-import { CommandDefinition } from "../../CommandDefinition";
+import { CommandDefinition } from "../../types/CommandDefinition";
 import { Category } from "../../constants";
 
 export const stop: CommandDefinition = {
